requirements.txt
streamlit
python-dotenv
Pillow
google-generativeai
pandas
promptwizard 